ast: move the traits into their own module
